% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_urls}
\alias{classify_urls}
\title{Take vector of URLs and classify into URL types}
\usage{
classify_urls(urls)
}
\arguments{
\item{urls}{Vector with URLs.}
}
\description{
This function takes a vector of URLs and will use a dictionary
of domains to classify into different types: survey platforms,
mail, google, social media, etc. and most importantly, news sites.

Eventually, we will have an argument that will contain a dictionary of
URL domains and their corresponding type. For now, everything is
just hard-coded in the function.
It will return a data frame with one row per URL and Boolean variables 
indicating whether the URL corresponds to each type.
}
\author{
Anthony Ramos, Simon Munzert, Pablo Barbera
}
