% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_urls_new}
\alias{classify_urls_new}
\title{classify_urls_new}
\usage{
classify_urls_new(urls)
}
\arguments{
\item{url}{vector of urls}
}
\description{
takes a vector of URLs, parses them, and classifies them by category
}
