% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_libraries.R
\name{discover_libraries}
\alias{discover_libraries}
\title{Discover R libraries used in project}
\usage{
discover_libraries(
  path = ".",
  pattern = "\\\\.Rmd$|\\\\.qmd$|\\\\.r$|\\\\.R$",
  recursive = TRUE
)
}
\arguments{
\item{path}{the root of the project to look for}

\item{pattern}{the file types to analyse (default: .R, .r, .qmd, .Rmd)}

\item{recursive}{search files recursively (default: TRUE)}
}
\value{
list of library names
}
\description{
Inspired by \href{https://gist.github.com/debruine/55bc4a23e1ab9d378e8cdcd3e9cabaf5}{Lisa DeBruine's gist}
}
\examples{
\dontrun{
find_libraries(".")
}
}
\seealso{
Other libraries: 
\code{\link{list_missing_libraries}()}
}
\concept{libraries}
