% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{download_in_parallel}
\alias{download_in_parallel}
\title{Parallel Download URLs}
\usage{
download_in_parallel(urls)
}
\arguments{
\item{a}{vector of URLs}
}
\description{
removes a vector of URLS in parallel, by default uses 2 cores for a dual-core processor
}
