% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{download_url}
\alias{download_url}
\title{Download URLs}
\usage{
download_url(url, path)
}
\arguments{
\item{url}{string of url, string of file name}

\item{path}{Path where html file will be stored}
}
\description{
downloads the HTML of URLs into a given file name
}
