% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankFurt.R
\name{getFrankfurtHeadlines}
\alias{getFrankfurtHeadlines}
\alias{getFrankfurtArticle}
\title{This package is for scraping the Zeit online front page}
\usage{
getFrankfurtHeadlines()

getFrankfurtArticle(url)
}
\arguments{
\item{url}{string}
}
\value{
dataframe with date, headline, summary and maintext as columns
}
\description{
Scrapes front page headlines from Frankfurt Allgemaine

scrapes the date, summary, title, and main text from an article on Frankfurt Allgemaine
}
