% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unix_time.R
\name{fromPosixTime}
\alias{fromPosixTime}
\title{Convert a Unix/Posix time to a timestamp}
\usage{
fromPosixTime(posix)
}
\arguments{
\item{posix}{a vector of integer numbers of seconds since Unix Epoch}
}
\value{
a POSIXct object
}
\description{
Convert a Unix/Posix time to a timestamp
}
\examples{
fromPosixTime(1501668000)

}
