% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unix_time.R
\name{fromUnixTime}
\alias{fromUnixTime}
\title{Convert a Unix time to a timestamp}
\usage{
fromUnixTime(posix)
}
\arguments{
\item{posix}{a vector of integer number of seconds since Unix Epoch}
}
\value{
a POSIXct object
}
\description{
Convert a Unix time to a timestamp
}
\examples{
fromUnixTime(1501668000)

}
\seealso{
fromPosixTime
}
