% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\name{getAlexa}
\alias{getAlexa}
\title{getAlexa}
\usage{
getAlexa(path = "Top/News")
}
\arguments{
\item{path}{path to be scraped}
}
\description{
Uses both the aws.alexa api and web scraping to get news categories for websites
}
