% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankFurt.R
\name{getFrankfurtHeadlines}
\alias{getFrankfurtHeadlines}
\title{Scrape Headlines from Frankfurt Allgemaine}
\usage{
getFrankfurtHeadlines()
}
\description{
Scrapes front page headlines from Frankfurt Allgemaine
}
