% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_libraries.R
\name{list_missing_libraries}
\alias{list_missing_libraries}
\title{List libraries not installed}
\usage{
list_missing_libraries(libs)
}
\arguments{
\item{libs}{name of libraries}
}
\value{
names of libraries not installed in current system
}
\description{
List libraries not installed
}
\examples{
\dontrun{
discover_libraries(".") |>
  list_missing_libraries()
}
}
\seealso{
Other libraries: 
\code{\link{discover_libraries}()}
}
\concept{libraries}
