% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_urls}
\alias{parse_urls}
\alias{download_url}
\alias{download_in_parallel}
\alias{parse_html}
\title{package containing parsing and downloading functions}
\usage{
parse_urls(urls, url_df)

download_url(url)

download_in_parallel(urls)

parse_html(filename, output)
}
\arguments{
\item{vector}{of urls, and dataframe to put results in.}

\item{string}{of url, string of file name}

\item{a}{vector of URLs}

\item{string}{of name of file to download, string of output file}
}
\description{
Parse a vector of URLs, and put them into a dataframe.

downloads the HTML of URLs into a given file name

removes a vector of URLS in parallel, by default uses 2 cores for a dual-core processor

takes in a file name to download, and downloads it to output
}
