% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_html}
\alias{parse_html}
\title{HTML Article Extraction}
\usage{
parse_html(filename, output)
}
\arguments{
\item{string}{of name of file to download, string of output file}
}
\description{
takes in a file name to download, and downloads it to output
}
