% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_urls}
\alias{parse_urls}
\title{Parse URLs}
\usage{
parse_urls(urls, url_df)
}
\arguments{
\item{vector}{of urls, and dataframe to put results in.}
}
\description{
Parse a vector of URLs, and put them into a dataframe.
}
