% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{retrieve_data_from_db}
\alias{retrieve_data_from_db}
\title{Template for retrieving data from PRU databases}
\usage{
retrieve_data_from_db(wef, til)
}
\arguments{
\item{wef}{start date[time], i.e. '2018-01-23 15:34'}

\item{til}{end date[time] (excluded)}
}
\value{
a tibble of data
}
\description{
You have to look at the code, save it as your own function with your own
SQL query and details.
It uses the following packages:
* `ROracle` (and hence `DBI`);
* `parsedate` as an example to parse input datetime params
* `withr` to manage locally set environment variables
* `tibble` to return a nice dataframe
Change according to your needs.
}
\examples{
\dontrun{
retrieve_data_from_db('2018-01-23 15:34', '2018-01-25')
}
}
