% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-package.R
\docType{package}
\name{scrap-package}
\alias{scrap}
\alias{scrap-package}
\title{scrap: A collection of useful (for me) snippets of code}
\description{
A collection of functions and snippets of code that I found useful in my data analysis in personal or professional projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/espinielli/scrap}
  \item \url{https://espinielli.github.io/scrap/}
  \item Report bugs at \url{https://github.com/espinielli/scrap/issues}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@gmail.com}

}
\keyword{internal}
