% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRAP-package.R
\docType{package}
\name{SCRAP-package}
\alias{SCRAP-package}
\alias{SCRAP}
\title{Scraping function for Paying Attention to 
Attention research project}
\description{
This package provides a series of functions that allow R users
to access Facebook's API to get information about users and posts,
and collect public status updates that mention specific keywords.
}
\author{
Anthony Ramos, Simon Munzert, Pablo Barbera
}
