% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrape.ZeitHeadlines}
\alias{scrape.ZeitHeadlines}
\title{Packages for scraping articles from Zeit Online}
\usage{
scrape.ZeitHeadlines()
}
\description{
This function takes the headlines off of ZeitOnline and returns a dataFrame with two
columns, titles and links.
}
\author{
Anthony Ramos, Simon Munzert, Pablo Barbera
}
