% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Faz.R
\name{scrapeFazArticle}
\alias{scrapeFazArticle}
\title{Scrape individual articles from the Frankfurter Allgemeine website}
\usage{
scrapeFazArticle(url, path)
}
\arguments{
\item{url}{string containing URL of article from Frankfurter Allgemeine}

\item{path}{Path where file will be stored}
}
\description{
This function takes a URL from an article from Frankfurter Allgemeine, and returns
a data frame with the URL, date, title, comments, summary, and text.
}
\details{
Return values will eventually contain main text
}
