% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Faz.R
\name{scrapeFazHeadlines}
\alias{scrapeFazHeadlines}
\title{Scrape homepage of Frankfurter Allgemeine}
\usage{
scrapeFazHeadlines(path)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
This function takes the headlines off of Frankfurter Allgemeine and returns a dataFrame with two
columns: titles and URLs.
}
\author{
Simon Munzert, Pablo Barbera, Joshua Timm
}
