% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreitagOnline.R
\name{scrapeFreitagArticle}
\alias{scrapeFreitagArticle}
\title{scrapes an article off of freitag.de}
\usage{
scrapeFreitagArticle(url, path)
}
\arguments{
\item{url}{string of article url}

\item{path}{Path where file will be stored}
}
\description{
Scrapes article date, content, and summary from freitag article
}
