% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreitagOnline.R
\name{scrapeFreitagHeadlines}
\alias{scrapeFreitagHeadlines}
\title{Scrape homepage of Freitag}
\usage{
scrapeFreitagHeadlines(path)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
This function takes the headlines off of Freitag and returns a dataFrame with two
columns: titles and URLs.
}
\author{
Simon Munzert, Pablo Barbera, Joshua Timm
}
