% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreitagOnline.R
\name{scrapeFreitagRSS}
\alias{scrapeFreitagRSS}
\title{Scrape RSS feeds of Freitag}
\usage{
scrapeFreitagRSS(folder)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
scrapes RSS feeds of Freitag home page
}
