% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmxOnline.R
\name{scrapegmxArticle}
\alias{scrapegmxArticle}
\title{scrapes an article off of gmx.de}
\usage{
scrapegmxArticle(url, path)
}
\arguments{
\item{url}{string of article url}

\item{path}{Path where file will be stored}
}
\description{
Scrapes article date, content, and summary from GMX article
}
