% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmxOnline.R
\name{scrapegmxHeadlines}
\alias{scrapegmxHeadlines}
\title{Scrape homepage of GMX}
\usage{
scrapegmxHeadlines(path)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
This function takes the headlines off of GMX and returns a dataFrame with two
columns: titles and URLs.
}
\author{
Simon Munzert, Pablo Barbera, Joshua Timm
}
