% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmxOnline.R
\name{scrapegmxRSS}
\alias{scrapegmxRSS}
\title{Scrape RSS feeds of GMX}
\usage{
scrapegmxRSS(folder)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
scrapes RSS feeds of GMX home page
}
