% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JungefreiheitOnline.R
\name{scrapeJungefreiheitArticle}
\alias{scrapeJungefreiheitArticle}
\title{scrapes an article off of Jungefreiheit.de}
\usage{
scrapeJungefreiheitArticle(url, path)
}
\arguments{
\item{url}{string of article url}

\item{path}{Path where file will be stored}
}
\description{
Scrapes article date, content, and summary from Jungefreiheit article
}
