% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JungefreiheitOnline.R
\name{scrapeJungefreiheitHeadlines}
\alias{scrapeJungefreiheitHeadlines}
\title{Scrape homepage of Jungefreiheit}
\usage{
scrapeJungefreiheitHeadlines(path)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
This function takes the headlines off of Jungefreiheit and returns a dataFrame with two
columns: titles and URLs.
}
\author{
Simon Munzert, Pablo Barbera, Joshua Timm
}
