% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JungefreiheitOnline.R
\name{scrapeJungefreiheitRSS}
\alias{scrapeJungefreiheitRSS}
\title{Scrape RSS feeds of Jungefreiheit}
\usage{
scrapeJungefreiheitRSS(folder)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
scrapes RSS feeds of Jungefreiheit home page
}
