% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpiegelOnline.R
\name{scrapeSpiegelArticle}
\alias{scrapeSpiegelArticle}
\title{scrapes an article off of spiegel.com}
\usage{
scrapeSpiegelArticle(url, path)
}
\arguments{
\item{url}{string of article url}

\item{path}{Path where file will be stored}
}
\description{
Scrapes article date, content, and summary from spiegel article
}
