% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpiegelOnline.R
\name{scrapeSpiegelRSS}
\alias{scrapeSpiegelRSS}
\title{Scrape RSS feeds of Spiegel Online}
\usage{
scrapeSpiegelRSS(folder)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
scrapes RSS feeds of of Spiegel Online home page
}
