% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpiegelOnline.R
\name{scrapeSpiegelRSSarticles}
\alias{scrapeSpiegelRSSarticles}
\title{Scrape articles from RSS feeds of Spiegel Online}
\usage{
scrapeSpiegelRSSarticles(folderInput, folderOutput, donefile)
}
\arguments{
\item{folderInput}{Folder where files with RSS feeds are located}

\item{folderOutput}{Folder where articles in html format will be stored}

\item{donefile}{File with URLs that have already been downloaded}
}
\description{
scrapes articles from RSS feeds of Spiegel Online home page
}
