% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrapeZeitArticle}
\alias{scrapeZeitArticle}
\title{Scrape individual articles from the Zeit Online website}
\usage{
scrapeZeitArticle(url, path)
}
\arguments{
\item{url}{string containing URL of article from ZeitOnline}

\item{path}{Path where file will be stored}
}
\description{
This function takes a URL from an article from Zeit Online, and returns
a data frame with the URL, date, title, comments, summary, and text.
}
\details{
Return values will eventually contain main text
}
