% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrapeZeitArticles}
\alias{scrapeZeitArticles}
\title{scrapeZeitArticles}
\usage{
scrapeZeitArticles(url)
}
\arguments{
\item{string}{containing URL of article from ZeitOnline}
}
\description{
This function takes the title, summary, and main text from an article from Zeit Online, and returns
a vector with the date,title, and summary.
}
\details{
Return values will eventually contain main text
}
