% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrapeZeitHeadlines}
\alias{scrapeZeitHeadlines}
\title{Scrape homepage of Zeit Online}
\usage{
scrapeZeitHeadlines(path)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
This function takes the headlines off of ZeitOnline and returns a dataFrame with two
columns: titles and URLs.
}
\author{
Anthony Ramos, Simon Munzert, Pablo Barbera
}
