% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrapeZeitRSS}
\alias{scrapeZeitRSS}
\title{Scrape RSS feeds of Zeit}
\usage{
scrapeZeitRSS(folder)
}
\arguments{
\item{path}{Path where file with homepage in html format will be stored}
}
\description{
scrapes RSS feeds of Zeit home page
}
