% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrapeZeitRSSarticles}
\alias{scrapeZeitRSSarticles}
\title{Scrape articles from RSS feeds of Zeit}
\usage{
scrapeZeitRSSarticles(folderInput, folderOutput, donefile)
}
\arguments{
\item{folderInput}{Folder where files with RSS feeds are located}

\item{folderOutput}{Folder where articles in html format will be stored}

\item{donefile}{File with URLs that have already been downloaded}
}
\description{
scrapes articles from RSS feeds of Zeit home page
}
