% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpiegelOnline.R
\name{scrape.SpiegelOnline}
\alias{scrape.SpiegelOnline}
\alias{scrape.SpiegelArticle}
\title{Package for Scraping articles off of spiegel online}
\usage{
scrape.SpiegelOnline()

scrape.SpiegelArticle(url)
}
\arguments{
\item{string}{containing url of function}
}
\description{
scrapes headlines off of spiegel online home page

Scrapes article date, content, and summary from spiegel article
}
