% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrape.ZeitHeadlines}
\alias{scrape.ZeitHeadlines}
\alias{scrapeZeitArticles}
\title{Packages for scraping articles from Zeit Online}
\usage{
scrape.ZeitHeadlines()

scrapeZeitArticles(url)
}
\arguments{
\item{string}{containing URL of article from ZeitOnline}
}
\description{
This function takes the headlines off of ZeitOnline and returns a dataFrame with two
columns, titles and links.

This function takes the title, summary, and main text from an article from Zeit Online, and returns
a vector with the date,title, and summary.
}
\details{
Return values will eventually contain main text
}
\author{
Anthony Ramos, Simon Munzert, Pablo Barbera
}
