% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpiegelOnline.R
\name{scrapeSpiegelArticle}
\alias{scrapeSpiegelArticle}
\title{scrapes an article off of spiegel.com}
\usage{
scrapeSpiegelArticle(url)
}
\arguments{
\item{url}{string of article url}
}
\description{
Scrapes article date, content, and summary from spiegel article
}
