% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitOnline.R
\name{scrapeZeitArticles}
\alias{scrapeZeitArticles}
\title{Scrape individual articles from the Zeit Online website}
\usage{
scrapeZeitArticles(url)
}
\arguments{
\item{url}{string containing URL of article from ZeitOnline}
}
\description{
This function takes a URL from an article from Zeit Online, and returns
a data frame with the URL, date, title, comments, summary, and text.
}
\details{
Return values will eventually contain main text
}
