% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitignore.R
\name{use_gitignore}
\alias{use_gitignore}
\title{Create a .gitignore for R projects}
\usage{
use_gitignore()
}
\description{
The content of the .gitignore is from
https://github.com/github/gitignore/blob/master/R.gitignore
}
\examples{
\dontrun{
use_gitignore()
}
}
